+++ tests/components/relatedProductsTests/OutfitCard.test.js
@@ -7,7 +7,7 @@ import 'regenerator-runtime/runtime'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
 
-import OutfitCard from '../../../client/src/components/userOutfits/OutfitCard.jsx'
+import OutfitCard from '../../../client/src/components/relatedProducts/OutfitCard.jsx'
 
 describe('Outfit Card', () => {
   test('renders category, name, and price in outfit card', () => {
